+++ cryptodb/order.go
@@ -7,8 +7,8 @@ import (
 )
 
 type Order struct {
-	ID            uint
-	PlanID        uint            `gorm:"index;not null"`
+	ID            uint64
+	PlanID        uint64          `gorm:"index;not null"`
 	Status        Status          `gorm:"type:varchar(25)"` // TODO: investigate option to define enums, see: https://github.com/go-gorm/gorm/issues/1978#issuecomment-476673540
 	OrderKind     OrderKind       `gorm:"type:varchar(25)"`
 	Size          decimal.Decimal `gorm:"type:decimal(20, 8)" json:"qty"`
@@ -21,7 +21,7 @@ type Order struct {
 
 const MaxTakeProfits = 5
 
-func NewOrders(PlanID uint) []Order {
+func NewOrders(PlanID uint64) []Order {
 	return []Order{
 		{PlanID: PlanID, Status: Unplanned, OrderKind: MarketStopLoss},
 		{PlanID: PlanID, Status: Unplanned, OrderKind: LimitStopLoss},
