+++ exchange/position.go
@@ -11,13 +11,13 @@ type Position struct {
 }
 
 type PositionResponse struct {
-	ReturnCode       int        `json:"ret_code"`
+	ReturnCode       int64      `json:"ret_code"`
 	ReturnMessage    string     `json:"ret_msg"`
 	ExtendedCode     string     `json:"ext_code"`
 	Positions        []Position `json:"result"`
 	ExtendedInfo     string     `json:"ext_info"`
 	ServerTime       string     `json:"time_now,string"`
-	RateLimitStatus  int        `json:"rate_limit_status"`
-	RateLimitResetMS int        `json:"rate_limit_reset_ms"`
-	RateLimit        int        `json:"rate_limit"`
+	RateLimitStatus  int64      `json:"rate_limit_status"`
+	RateLimitResetMS int64      `json:"rate_limit_reset_ms"`
+	RateLimit        int64      `json:"rate_limit"`
 }
